Remove obsolete combined transaction tests
