+++ tests/helpers.py
@@ -1,0 +1,30 @@
+import os
+import sys
+import tempfile
+from pathlib import Path
+
+from sqlalchemy import create_engine
+from sqlalchemy.orm import sessionmaker
+
+# Ensure the project root is on the Python path
+sys.path.append(str(Path(__file__).resolve().parents[1]))
+
+from budget import database
+
+
+def get_temp_session():
+    db_fd, db_path = tempfile.mkstemp()
+    os.close(db_fd)
+    engine = create_engine(f"sqlite:///{db_path}", future=True)
+    TestingSession = sessionmaker(bind=engine)
+    database.Base.metadata.create_all(engine)
+    return TestingSession, Path(db_path)
+
+
+def make_prompt(responses):
+    iterator = iter(responses)
+
+    def _prompt(*args, **kwargs):
+        return next(iterator)
+
+    return _prompt
